- Fix UserService check for original user - Move dao classes to separate modules
