+++ compass/views/api/photo.py
@@ -1,6 +1,6 @@
 # Copyright 2022 UW-IT, University of Washington
 # SPDX-License-Identifier: Apache-2.0
-from compass.dao import PhotoDAO
+from compass.dao.photo import PhotoDAO
 from datetime import datetime, timedelta
 from django.core.exceptions import ObjectDoesNotExist
 from django.http import HttpResponse, StreamingHttpResponse
