+++ docker/settings.py
@@ -83,6 +83,6 @@ GOOGLE_ANALYTICS_KEY = os.getenv("GOOGLE_ANALYTICS_KEY", default="")
 SUPPORTTOOLS_PARENT_APP = "Compass"
 SUPPORTTOOLS_PARENT_APP_URL = "/"
 
-USERSERVICE_VALIDATION_MODULE = 'compass.dao.UserServiceDAO.is_netid'
+USERSERVICE_VALIDATION_MODULE = 'compass.dao.person.is_netid'
 USERSERVICE_OVERRIDE_AUTH_MODULE = 'compass.views.can_override_user'
 RESTCLIENTS_ADMIN_AUTH_MODULE = 'compass.views.can_proxy_restclient'
